+++ tests/testgame.cpp
@@ -82,8 +82,8 @@ void TestGame::OnEvent(Event &evt)
                 double len = std::sqrt(pos[0]*pos[0] + pos[1]*pos[1]);
                 theta += dtheta;
                 glm::vec3 newpos = pos;
-                newpos[0] = len * std::cos(theta);
-                newpos[1] = len * std::sin(theta);
+                newpos[0] = static_cast<float>(len * std::cos(theta));
+                newpos[1] = static_cast<float>(len * std::sin(theta));
                 _camera.SetPosition(newpos);
             }
             break;
@@ -114,7 +114,7 @@ void TestGame::OnEvent(Event &evt)
             const auto& lookAt = _camera.GetLookAt();
             const auto& eye    = _camera.GetPosition();
             glm::vec3 dir = eye - lookAt;
-            float scale = 0.1f * mouseEvt.YOffset();
+            float scale = static_cast<float>(0.1 * mouseEvt.YOffset());
             dir = dir +  dir * scale;
             _camera.SetPosition(lookAt + dir);
             break;
@@ -169,7 +169,7 @@ void TestGame::SetupTerrain()
     std::vector<unsigned int> conn = mbos::ibo(_mesh, flatShading);
 
     _vao = std::make_unique<VertexArray>();
-    _vbo = std::make_unique<VertexBuffer>(vertices.data(), vertices.size() * sizeof(float));
+    _vbo = std::make_unique<VertexBuffer>( vertices.data(), static_cast<unsigned int>(vertices.size() * sizeof( float )) );
 
     VertexBufferLayout layout;
     layout.Push<float>(3); // position
@@ -177,7 +177,7 @@ void TestGame::SetupTerrain()
     layout.Push<float>(2); // texture coordinate
     _vao->AddBuffer(*_vbo, layout);
 
-    _ibo = std::make_unique<IndexBuffer>(conn.data(), conn.size());
+    _ibo = std::make_unique<IndexBuffer>(conn.data(), static_cast<unsigned int>(conn.size()));
 
     _shader = std::make_unique<Shader>("res/shaders/basic.shader");
 }
